[Sema] splitted source file
